fix: token表示サンプルページの修正
