+++ packages/mottai_flutter_app/lib/development/development_items/ui/development_items.dart
@@ -14,7 +14,7 @@ import '../../../job/ui/job_update.dart';
 import '../../../map/ui/map.dart';
 import '../../../worker/ui/worker.dart';
 import '../../color/ui/color.dart';
-import '../../fcm_token/fcm_token.dart';
+import '../../fcm_token/ui/fcm_token.dart';
 import '../../firebase_messaging/ui/firebase_messaging.dart';
 import '../../firebase_storage/ui/firebase_storage.dart';
 import '../../force_update/ui/force_update.dart';
